requirements.txt
pyyaml>=6.0.1
watchdog>=3.0.0
python-frontmatter>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.1.0
regex>=2023.10.3
pytest>=7.4.3
psutil
